Change generated gwblocks.pvlist file to be more selective
The generated gwblock.pvlist used wildcards and so could match
name queries from other instrumemnts that got forwarded on.
This was not a real problem when the block gateway was purely
on localhost network, but as it can now resolve off instrumement
e.g. for beam currrent it needs to be much more selective